Manage /etc/crontab and splay times across boxes
The /etc/crontab file controls the timings of the periodic jobs.

The ubuntu defaults are:

 - hourly: 17 minutes past the hour
 - daily: 0625
 - weekly: 0647
 - monthly: 0642

Unfortunately that means all machines will run their jobs at the same
time, which is not desirable for performance reasons.

By using fdqn_rand() we can choose a value which is randomised across
machines but it related to the hostname so consistent across puppet
runs (and indeed machine rebuilds). Under the hood, fqdn_rand uses
the puppet
[deterministic_rand](https://github.com/puppetlabs/puppet/blob/master/lib/puppet/util.rb#L503-L513)
function.