Moving our jail config to jail.local
This is the first step - these were the changes that had already been made when
the module was initially added: https://github.gds/gds/puppet/commit/e343f1a

The only one I have not included is:

ignoreip = 127.0.0.1/8

Our version was:

ignoreip = 127.0.0.1

The default is more correct, so I have left it.

As per previous commit, I have not changed back to the default jail.conf; I will
remove this in a later commit as it is not required.