Apply other changes to jail.local
Since it is reasonably likely that we will remove this module, I have referenced
the git commits as a comment by each change, for auditing.

I did not include this change: https://github.gds/gds/puppet/commit/e3d9be
because that line is not missing from the default jail.conf - i.e. if we had
done it correctly initially we might not have had that bug.