[#58233642] Collect Fastly stats into Graphite
As an Operator I want Fastly stats in Graphite so that we can overlay
them, with other metrics that we collect in order to correlate events.

1. Installs the statsd-ruby gem which is an easy interface to submitting
StatsD metrics.
2. Adds a scripts that will take the last 5 minutes of Fastly stats for a
particular service and send statsd gauges of the format:
```
fastly.<servicename>.5min_avg.<stat>:<value>|g
```
3. Configures a cron task (running every minute) per Fastly service that
will submit the stats to the local StatsD port (8125/udp).