[#58269310] Create govuk::app::router for s_cache
In some respects it's overkill to use `govuk::app` and `Procfile` for such a
simple app. However it gives us a lot of sensible defaults, like logging and
monitoring.

We call `govuk::app::envvar` to setup some move vars that are specific to
the router. These are intentionally namespaced so that the router can be
used outside of this configuration, especially by users of the open source
repo.

A list of MongoDB nodes are passed in from `s_cache`. It seems safer to do
this than make them dymanic, at the risk of restarting the router
frequently. They could be moved to extdata.

Ports are taken from gds/development@fa6fd52 and the Varnish config will
need updating respectively. `Procfile` will need adding to the router repo.
The public (reverse proxy part) intentionally only binds to loopback.