[#58269310] Process stats for app_type => bare
Use `$command` to create a regex for collectd to get process stats. This
previously assumed that all apps were Rack/Unicorn. This will remove
(broken) monitoring for `app_type => procfile`.

I'm not 100% sure that the regex will always work. Or capture all of the
necessary child processes. We should keep an eye on that.