Feature flag Nagios checks for Mirror
These currently only work on in production because of firewalling on the
Mirror environment. The checks are marked as acknowledged in all other
environments.

Patch this broken window by disabling the alerts in other environments. We
only need to be alerted by prod. Enabled for prod in gds/deployment@8a690b8