Tidy up envvar resources for Router
These were incorrectly cribbed from `govuk::app::config`. The `${title}-`
prefix was evaluating to the name of the class `govuk::apps::router`.

We don't need to do this, since the environment variable names themselves
are namespaced. The Puppet resource title won't clash with anything else. So
we can omit the prefix and `varname` will default to the `$title` within
`govuk::app::envvar`.

Noop on an existing machine.