[#58880904] Safety check for Hiera
It's possible that if someone pulls the `puppet` repo but doesn't pull
`vagrant-govuk` or `vagrant reload` their VM then Puppet won't be able to
load data from Hiera. Because we now expect APT configuration to come from
there, this can result in some silent and rather unexpected results.

Add a key to `hieradata` and a condition to `site.pp` which tests that Hiera
has been configured correctly and is working. Fail loudly if it's not able
to find that key, suggesting that something is wrong.