Document the fact we are on HSTS preload lists
I suprised Nick Stenning when he mentioned getting on the HSTS Preload
List for Chrome, by telling him I already did it in Nov 2012.

I then checked my mail and it appears that the only record I have of
this is an email conversation between myself, Etienne, JamesS and Adam
Langley from Google. There's no record that I broadcasted it more
widely.

I can't think of a better place to document that we are on the HSTS list
appart from where we assert the HSTS headers for GOV.UK - so here we
are.