[#60420920] Feature flag Router redirects
New feature flag to allow the Router to serve redirects instead of handling
them in Nginx above. Needs to be used together with the Varnish feature
flag, which is enabled in preview, otherwise the Router won't see the
request.

All this code can be removed a short while after launch.