Refactor VCL condition for Licensing cookies
Split the two conditionals that treat Licensing as a special case:

- Cookies should be removed from all applications except for Licensing.
- Responses from Licensing should have a header injected to preserve it's
  own error pages.

This makes the condition for the first case syntactically the same in
`vcl_recv{}` as `vcl_fetch{}`. Which should help to prevent my copy/paste
mistake that occurred in bd1a957 and fa6a420. It should also be easier to
read, since these two conditionals aren't strictly related at all.

Sidenotes:

- Tidied up comments.
- Camel-cased the `beresp.http.set-cookie` HTTP header name. Varnish
  evalutes this as case-insensitive, per the RFC, but this makes it
  consistent with the other `Set-Cookie` reference.
- Includes a note (paraphrased from someone else) about why we are stripping
  cookies in the first place.