Manage automysqlbackup permissions using module
At the moment, we attempt to manage permissions using a postbackup
script; however, the automysqlbackup::backup type also managed
permissions and resets them to its default values of 0700 (directories)
and 0400 (files) after we set them up in the postbackup. This means that
the govuk-backup user cannot later read them.

If the module can manage the backup file permissions, we should probably
leave it to the module rather than trying to manage it ourselves.