Remove conditionals for $::govuk_provider
We're no longer actively running or deploying to a Puppet Master in EC2. So
we can remove the conditionals that supported the transition from EC2 to
Skyscape (interim).

None of these should affect the development VM which doesn't call
`mongodb::configure_replset`, pass it's own `FACTER_govuk_provider`, or
inherit from `s_base`.