Update mongodb class to take a version (take 2)
This will allow us to have servers with different versions of mongo
installed. This also means that the version installed is consistent. At
the moment, the class just does ensure => installed against the
mongodb20-10gen package, which means that the version installed will
depend on when the server was provisioned. Most servers have 2.0.7
installed, whereas the router-backend machines have 2.0.8 installed.

This has required adding another file to the hiera hierarchy so that we
can set versions on a per-class basis.