Warn hiera.yaml changes require master restart
Discovered when the addition of `%{::environment}` to the hierachy in
40a72a0 wasn't picked up by 63f3386, causing lines to be removed from the
config, and required reverting in b3b2e3c.

The file is just pushed to disk using Capistrano like the rest of the Puppet
modules/manifests. We can't trigger a reload of `puppet::master::service`
because the `hiera.yml` isn't managed by Puppet - doing so would mean a
slow two-phase deployment.