Surface Whitehall overdue healtcheck URL in Nagios
Make the URL that `check_http` calls available in the Nagios UI and alerts.
Will make it easier for 2nd line to check the current status, without
needing to necessarily log into a machine and run the command in the
opsmanual. Requires BasicAuth.

It would be nice if `check_http` could just report the returned and expected
strings. But alas not.