Reduce Nagios/Graphite graphs to 2hrs
The default of 24hrs is too long. It's very hard to see an emerging issue on
the right hand of the graph. So reduce the time period to the last 2 hours.

By including the param in the URL it's also easy to adjust by hand once you
have it up in a browser.