[#61198890] Disable GC alerts for logs-es
These alerts are high value on our main Elasticsearch cluster because they
indicate that applications like Rummager, and that apps that depend on that,
are going to choke.

However they are of very little value on our `log-elasticsearch` cluster
which is used by Kibana. We know it contains a lot of data, and that some
queries are slow, but it's usually obvious to the person running the search.
We don't need to bother 2nd line with that information.

No spec tests because we can't test exported resources. We'll need to purge
the existing checks from PuppetDB/Nagios.