[#51136581] Use HTTPS for frontend->backend apps
Use HTTPS for all frontend->backend communication. And frontend->frontend,
in the case of static->whitehall.

I don't necessarily agree with using HTTPS everywhere internally - because
it's slow to terminate and re-encrypt on every hop, and we're using the same
cert/key pair everywhere.

However this does make it consistent throughout and will allow us to disable
HTTP for public facing vhosts on `backend-lb`. When deployed we should be
able to change `loadbalance::balance` resources to `https_only => true`.

To identify requests currently affected:
```
@source_host:backend-lb-* AND _exists_:@fields.request AND @fields.server_port:80
```

For a request to `https://www.gov.uk/government/uploads/foo` it travels:

- www.gov.uk:443 (cache)
- whitehall-frontend.production.alphagov.co.uk:80 (frontend-lb) [1]
- whitehall-frontend.production.alphagov.co.uk:443 (whitehall-frontend)
- whitehall-admin.production.alphagov.co.uk:80 (backend-lb) [2]
- whitehall-admin.production.alphagov.co.uk:443 (backend)

[1] Because Varnish/Router can't speak HTTPS.
[2] This is the one we're trying to fix.