[#51136581] Change loadbalancer https_only default
Change the default for `https_only` param of `loadbalancer::balance` to be
`true`. This will prevent us from creating HTTP vhosts on `backend-lb` where
the services behind should be HTTPS-only.

Existing behaviour is maintained using `https_only => false`. The ones
marked with `FIXME` comments can be removed and use the default of true when
8a55778 has been rolled out.

The behaviour of `frontend-lb` will not be changed because Varnish/Router
needs HTTP upstreams to reverse-proxy to.