[#51136581] HTTPS support for nginx::c::v::proxy
Add proper support for HTTPS upstreams in `nginx::config::vhost::proxy`.
This makes sure that both the ports in the `upstream{}` block and protocol
in the `proxy_pass` directive are correct. Which wasn't true of the previous
commit, where I changed the port in the `to` param but the `proxy_pass`
directives continued using `http://`.