higher rate limit for service feedback endpoint
one of the issues discovered during the recent quarterly GOV.UK pen test
was that it was possible to flood the support contacts database by submitting
many requests per second through the service feedback/report-a-problem POST endpoints.

This change is an attempt to mitigate this risk by introducing tighter rate limiting (6 req/min per IP)
for the service-feedback endpoint initially.