Redux: use statsd not logster data for nagios checks against graphite
Nagios checks which check graphite currently mostly use logster
data. This is bad for lots of reasons.

We tried to flip to use statsd data (under the stats. prefix) before,
but had to revert in 92d547573d5 because statsd doesn't create counters
until it receives data for the first time. For infrequently-used apps
such as designprinciples, they might never have issued a 5xx error, so
the 5xx counter may not exist at all.

However fc81fd74 changed the behaviour for missing metrics from CRITICAL
to UNKNOWN, which makes this less of an issue. We can therefore
reintroduce statsd metrics from nagios checks.

This doesn't affect varnish_logs metrics, which only have logster
data (because varnish logs are not in json event format and are not
parsed by tagalog).