[#61532234] Remove Nagios->Zendesk contact groups
These are a broken window. Nobody reviews Nagios alerts in Zendesk and as a
result most people have filtered all Zendesk notifications.

The `nagios_is_zendesk_enabled` feature flag should be renamed to reflect
what it's actually doing - which is enabling out-of-band notifications to
PagerDuty and Campfire. Separate commit/PR to follow.