[#61532234] Hiera for Nagios Campfire/PagerDuty
Move the configuration of Nagios contact groups for Campfire and PagerDuty
out to Hiera. This namespaces the keys correctly to this class, making it
more obvious where they are used, as well as renaming them to better reflect
their purpose.

Simplified the logic in the manifest by using separate variables and
puppetlabs-stdlib `flatten()` function to compose a list of lists. This
means we no longer have nested conditions.

The move to hiera and class params means that we can now test this as a
normal class. I've added contexts for three common use cases. There is some
missing coverage, such as absent Campfire credentials, but the tests are so
slow (3s per context) that I don't think there's value covering them all.

Depends on gds/deployment#90