Add memcached to backend and development nodes
There's a requirement for an ephemeral LRU-style cache, and it's simpler
for us to add memcached for backend machines to use than to fork our
Redis install into two separate configurations. Memcached offers us the
ephemeral cache characteristics we want out of the box and is very well
understood. It additionally integrates well with Rack::Cache and hence
our general stack.

This requirement was triggered by:

https://github.com/alphagov/govuk_content_api/pull/134

On the backend machines it's configured to use 12% of available RAM, and
in development it's strictly limited to 64MB. It's limited to listen on
loopback only on both machines.