Remove redundant location for Google verification.
A file corresponding to this does exist in the frontend application, but
nothing is currently routed to it (either on www or assets).  Even if
this route is needed, we don't want to be adding this unnecessary config
to _every_ application.

Note there is a different Google verification file that lives in static,
and is routed (static adds a route for it to the router).