Use transformNull() on Nagios statsd data
We went back to using statsd data for Nagios alerts in 363c7c5. But as noted
it will/is still producing lots of `UNKNOWN` alerts which, although lower
severity than before, are still distracting and confusing. Using the new
Fabric task to squash them one-by-one is too laborious.

As an alternative to reverting that commit (and losing the value of the
superior statsd data), we can resort to a hack where we munge `NULL` values
into `0`. This will prevent `UNKNOWN` alerts from firing for counters which
haven't been created yet (new or low traffic application, server rebooted).

This comes at the expense of not being able to tell if a metric genuinely
doesn't exist. We'll have to be more careful about typos.