Default the MySQL import user to root
Way back in the mists of time this needed to be configurable. However,
without a value (passing no -u command results in '@localhost' as the
user), the import blows up and won't complete.

We want this to succeed on development with minimal arguments, so we
should default the argument to root, as the root user has no password on
dev. On all other environments this might conceivably be run the root
user has a password so this should fail appropriately.