[#63060484] UNKNOWN for Whitehall pub timeouts
Wrap Nagios `check_http` with a simple script which converts `CRITICAL`
alerts for timeouts into `UNKNOWN`. Use this for the Whitehall overdue
publications alerts to reduce the possibility of false-positives when
Whitehall fails to serve the check endpoint. As seen recently over the
holiday period when everything had a bit of a wobbly.

This will presently still page on-call because the contact options include
UNKNOWN, WARNING and CRITICAL. But it should be clearer that it *might* be a
false-positive, or the side-effect of another problem, rather than simply a
publications problem.

As discussed in and supersedes #1204.