Allow multi-word room names for Campfire
This has been silently failing since I changed the room which Nagios
alerts get sent to. The cause of this is that the room name has changed
from one word (Zeitgeist) which does to require quoting to "GOV.UK
Monitoring Alerts" which would need to be quoted.

The addition of some quotes fixes this.