[#59908614] hosts(5) entry for Gor to staging
Create a `hosts(5)` entry for the hostname which Gor connects to staging.

This is in reaction to an incident yesterday, whereby Gor was making a DNS
request (or two -- IPv6) for `www-origin-staging.production` for every HTTP
request seen in production, causing other legit DNS requests to be rate
limited (Google DNS -- don't ask):

- https://docs.google.com/a/digital.cabinet-office.gov.uk/document/d/1voBAM5txYENkh6bkxUNmfL8Rdx9gJNOtPxV378hon_s/edit
- https://groups.google.com/a/digital.cabinet-office.gov.uk/forum/#!topic/2nd-line-support/qja3bRmkytM

Flagged by a class param which can be enabled only in production using
hieradata. This will be re-used to also setup the `puppet-gor` module
shortly.

We can subsequently remove the record from Dyn. Which is a good thing,
because it's a hack, and we can at least annotate it in here.