[#63456770] Don't guess Unicorn worker count
Rework the `check_unicorn_workers` Nagios script. Chiefly to not fallback to
a worker count of 4 if it's unable to determine the real value. This was
causing misleading Nagios alerts when the Ruby command failed. Instead
report UNKNOWN and indicate that it needs manual review.

I also took the opportunity to DRY up some of the code and be more
aggressive about catching errors.