Temp set num old indexes to 8
We have a problem with ES logging, in that the apps are using the
logstasher gem to send logs and one of the fields is a hash that causes
the index to be massive.

We've tried to fix that here: https://github.gds/gds/puppet/pull/1246
but it's not working. As per notes Alex is trying to fix the logstasher
gem, which will also resolve this issue. Until we come up with a
permanent fix we need to make sure we don't run out of space on the
logging machines.