[#59908614] Manage Gor for s_cache
Uses a vendored module to install Gor, create an upstart job, and start the
service. Our `govuk::gor` module is responsible for enabling or disabling
this based on certain feature flags. The service will be unconfigured
(targets removed) and stopped if neither feature flag is enabled.

The Gor arguments match what has been run manually on production in the past
week or so. They differ from the previous way that we used Gor, which was a
listener in production talking to a replayer in staging over an SSH tunnel.

This method of performing the listen and replay in production means that we
don't need to fiddle with SSH tunnels or VPNs, we just fire normal requests
to the staging origin, which are encrypted by HTTPS. This comes at the small
cost of increased load on the cache nodes -- which doesn't appear to be an
issue, but we should keep an eye on.