[#64033592] Move tika to a module
DRY up duplicate package resources by moving `tika` to a module and
including that from the two places that currently use it.

I'm not making this responsible for setting up Java. That should be explicit
in the node or app class that uses this, because you may want different
versions. The dependency on `java::set_defaults` will prevent tika from
being included without Java being setup.