Check if we are reaching our AWS quota limit
We should check if we are approaching our AWS quota. This does that.

Our quota is currently 1M/day. We usually use <10,000. So while a 30%
warning and a 50% critical seems low, it's actually 30-50x our expected
usage.

Also, this is going to Pagerduty people, because at the moment if it fires it
means we have a serious problem and we are halfway to breaking
Production.