[#64143224] Add an optional before to govuk::mount
Due to the unique way that hiera and create_resources work you cant just pass a require in your hiera file. To get round this add an option to govuk::mount to take the name of a govuk::lvm resource.

If this option is passed make sure that the govuk::lvm resource is run before the ext4mount.