Add ntp peer check
This adds a nagios check based on check_ntp_peer, which can interrogate
a running ntpd's state to determine if the server believes itself to be
in sync.

Unlike check_ntp_time, this alert doesn't contact an upstream ntp server
for comparison; it instead is entirely dependent on ntp's own
self-assessment, which you can see for yourself with `ntpq -p`. This
includes data based on ntp's estimated time offset relative to each of
its upstreams, jitter of the offset, etc etc.

It can also alert based on the number of upstream servers that the host
trusts ("truechimers") but check_ntp_peer only supports this option on
precise or later.

This does mean that we will be able to alert when NTP doesn't trust any
timesources, and fix the problem then, rather than waiting for the clock
to drift significantly before discovering that there's anything wrong.