Revert "reinstate check_ntp_time nagios command..."
This reverts commit cde5a08ca32c335d078c47306ed65e68ba2cfedf (yes, from
a year and a half ago).

The commit 81a2c2a removed the need for the check_ntp_time nagios
command and replaced it with an equivalent NRPE command. However, during
the transition phase while the storeconfigs caught up, nagios complained
that the check_ntp_time command was missing. cde5a08c reinstated the
command for the transition period and got left behind as cruft ever
since.