Use hiera to fetch app_domain instead of extlookup
gds/deployment#107 created an app_domain entry in hiera. This migrates
us to use it.

In order to ensure the tests continued to pass, I've added
hiera-puppet-helper, which allows you to specify a `let(:hiera_data)`
block for a test to stub out hiera data. There's no place to put global
hiera data; IMHO, this is a Good Thing.

There's a bit of nasty behaviour where if you don't specify all of the
parameters for a parameterized class, puppet will try to autoinject
values from hiera and blow up because of the way hiera-puppet-helper has
monkey-patched things, so I've added a counter-monkey-patch to just
completely disable automatic injection. This is horrible, and hopefully
we can find a better solution some time.

Once this is in, we can more easily gradually migrate one value at a
time from extdata to hiera.