Set nginx timeouts for app servers
Our application server nginx instances don't set timeouts, so they get
the default connect timeout of 60s and the default read timeout of
60s. This is Really Bad, particularly since the downstream proxies set
stricter timeouts (our load balancer nginx sets connect timeout 1s, read
timeout 15s; our cache box nginx sets read timeout 20s). This means that
even though the downstream has given up on the request, the appserver
nginx keeps trying and consumes a connection from a unicorn worker all
the way up to the 60s timeout.

This is a pretty dumb commit -- it just hardcodes to connect timeout 1s
and read timeout 15s. If we need configurability we can add it later.

The uses of proxy-vhost.conf are all via nginx::config::vhost::proxy,
which is used by:

* govuk::app::nginx_vhost, used by
  * govuk::app::config (used by most apps)
  * govuk::apps::bouncer
  * govuk::apps::whitehall
* govuk::apps::efg
* govuk::apps::publicapi
* govuk::node::s_graphite

I have slight concerns that graphite might not be able to cope, but
given that it's not user-facing I'm keen we just try it and see.

Almost all of the govuk::app instances are behind an nginx load balancer
instance, so there is no point at all in having timeouts higher than the
1s/15s set by the load balancer. Bouncer's fastly config has a connect
timeout of 1s, a first-byte timeout of 15s and a between-bytes timeout
of 10s. Not sure about EFG.