[#62158624] Make apt-add-repo work on Lucid
Omit the `-y` flag for Lucid because it's not supported -- it won't prompt
you by default. This allows us to provision/test new Lucid machines again,
which is handy while we still have the Interim platform. I need this for the
Lucid/Precise upgrade.

Besdies I hear Lucid is coming back into fashion!

:fax: :pager: :vhs: :minidisc: :radio: