[64143110] Rename Icinga SSL cert/key
This vhost is having a bit of an identity crisis. Most of it still refers to
"nagios", including the cert/key filenames.

On an upgraded machine that used to run Nagios (like Interim and Platform1)
this works because the old files still exist. But on a new machine Nginx
fails to start.

Rename them back to "nagios" for now to match the rest of the vhost.