Prefer SSH ClientAlive* to TCPKeepAlive
The keep-alive packets sent by the `TCPKeepAlive` SSH option may be
blocked by some firewalls. Furthermore, any blip in connectivity could
cause the sshd to drop the connection when `TCPKeepAlive` is enabled.

Prefer ClientAliveInterval and ClientAliveCountMax, which use the SSH
protocol itself to send keep-alives.

This should improve the reliability of SSH connections, for example to
prevent `broken pipe` errors in scripts.