[#57797586] Puppet never use cached catalog
Prevent Puppet from ever using a cached catalog. If the master is unable to
provide a catalog, due to a syntax or semantic errors in the manifests, then
a hard failure will occur. The NSCA check will alert us that it was unable
to parse the summary file. This improved alerting comes at the small expense
that existing resources will not be enforced (if they happen to be
out-of-sync) until the error is resolved.

This in turn reverts c4aa905 which extended the NSCA check to alert when a
cached catalog was used. A bit of digging it turns out that Puppet will
never write `last_run_report.yaml` if configured with `report = false`,
which explains 58ad3e3. We could enable `report`, which will also submit the
reports contents to the master (you can't separate the two features), but
even so `last_run_report.yaml` gets written *after* `postrun_command` is
executed. So it will always be one run behind.

The only way that we could improve on this is to enable reports and use a
custom report processor on the master to send the passive checks on behalf
of the clients. This would have the added benefit of having access to logs
from the report, so would could inject the actual errors into Nagios alert.
We should disable the `store` report if we do this because we don't want to
fill the disk with reports -- it might be good to store them in PuppetDB
though.