Include module_spec_helper and set mock framework
This is needed for testing Puppet functions. The `rspec-puppet` matchers
don't appear to be passing arguments correctly, and I can't see a way to
store the result in a variable, to compare it against multiple things.

Explicitly setting the mock framework so that `puppetlabs_spec_helper`
doesn't switch us back to `mocha` again and break the syntax in
`modules/govuk/spec/facter/fqdn_underscore_spec.rb`.