Link to Puppet logs in Nagios alerts
We frequently disregard alerts for "puppet last run errors" because it's too
much hassle to identify the source of the problem. Make it easier by linking
to a Kibana search that shows Puppet logs for the affected host.

This could be further refined to exclude certain log severities. But let's
see if that's necessary before doing it.

This is a budget solution to what I'd really like to do, which is to inject
the offending log lines into the passive check. But we can't do that without
writing a master-side report processor for Puppet per the comments in
63fe029.