Move puppet_passive_check_update to govuk_puppet
It turns out that `last_run_report.yaml` is written AFTER `postrun_command`
is executed. So our alert has always been one run (30mins) behind.

Fix this by moving it to our `govuk_puppet` wrapper script. Send STDOUT to
`/dev/null` so that it outputs errors but not:
```
1 data packet(s) sent to host successfully.
```

I did consider using a `trap` but I think it's the right behaviour for check
not to execute if the script is ^C-ed, in case the an old summary file is
used.

This does make it even more important that we use the `govuk_puppet` wrapper
and not the plain `puppet agent` command. Since we already manage
`/usr/bin/puppet` we might be able to prevent this -- but I'm not sure how
to identify that caller besides setting an environment variable.