[#65086976] Feat flag asset-{master,slave} rename
We can't change the hosts(5) entries on Skyscape Interim because the
ordering of the entry causes the output of `hostname -f` to change and bad
things happen, like new Graphite metrics are created.

So instead feature flag it with a class param. Default behaviour for
Interim. A new key in the `platform-one` branch of deployment with rename
Skyscape Platform1.