[#65205768] Kill PuppetDB Dashboard
TL;DR: The PuppetDB dashboard is currently accessible behind the VPN. Since we do not use it, this commit kills it's vhost. A caveat - that `puppetdb.cluster` needed to still work as `govuk_node_list` depends on it - has been addressed in order that this still works.

---

Long story:

Recently, it was noticed that the PuppetDB dashboard is accessible externally at http://puppetdb.production.alphagov.co.uk/. We have no need to use it, and therefore it was decided that the best course of action is to remove it. `govuk_node_list` relies on it being accessible, however, in order to work, and the vhost for `puppetdb.cluster` existed at both:

- modules/govuk/templates/node/s_deployment/jenkins.conf.erb
- modules/puppet/templates/puppetmaster-vhost.conf

This commit removes the vhost entirely from `jenkins.conf.erb` as it is surplus to requirements and allows me to meet the aim of the original story. This has been tested:

- puppetdb.production.alphagov.co.uk is no longer accessible
- `govuk_node_list` returns a list of nodes that Puppet is aware of, as expected
- puppetdb.cluster responds to ICMP requests