[#65294674] Hiera flags for SSH AllowUsers/X11
Use feature flags instead of `govuk_class` magick. I've separated the logic
for specifying the list of users and actually enabling the config option, so
that we can enable it for all jumpboxes in `hieradata/class` and then
provide environment specific user lists in `gds/deployment:puppet/hierdata`.

Data for production and staging added in gds/deployment#128