[#64241148] Move braces to encompass checks in if
This commit moves braces in order to encompass the Icinga checks in the if statement. If the value of `mirror::enable_checks` is true, then process the Icinga checks. The value defaults to false, and is only true in Production. Preview and Staging default to false, and are explicitly set to false in Hiera as a failsafe.

@jordanhatch noticed that alerts for the mirror being out of date were appearing in all three environments erroneously. This commit should fix that.