Remove whitehall-backend-lb
These machines were in Platform1, then we removed them. This puts the
removal back into master, to fix a difference in the platform-one
branch. The FIXME still applies - once we have switched platform-one to
master AND all machines are successfully running puppet (which isn't the
case at the moment) we can remove this section.