Move deploy ssh keys to Hiera
Once the correct keys are in the deployment hieradata, this change will
convert the keys we put into the deploy user's authorized_keys from
various different extdata variables to a single per-environment hash in
Hiera.

The reason I want this, is that I will be using different
jenkins_skyscape_production_key values in platform-one and interim and
these will be set from different branches of the deployment repo.