Move MongoDB version to deployment repo
Remove the line that sets the MongoDB version for the `router_backend`
nodes so that we can set it in the `deployment` repo.

This is so that we can synchronise the 'platform-one' branch with master
in this 'puppet' repo.

Platform 1 uses MongoDB version 2.4.8 by default; `router_backend` nodes
will inherit that version in Platform 1 but those on the Interim
platform will continue to use `2.0.8`.

Depends on gds/deployment#137.