Feature flag instead of $env for govuk mount/lvm
Use class params and hieradata to feature flag the non-creation of
`govuk::mount` and `govuk::lvm` resources on Vagrant VMs. Instead of
inspecting `$::environment` directly to figure out where we are running.

This is more flexible and transparent. It would be good to have a lint check
to prevent us ever looking up `$::environment` in the future.