nginx config to proxy some requests to UKBA
UKBA have some tools and transactions which they cannot move or replace in time
for the UKBA switchover.

This is an experiment to see if we can successfully proxy the appropriate
requests to them, but leave the bulk going to Bouncer to be redirected or
archived.

If this is successful, in the longer term we will likely add some features to
the router and use a dedicated instance of it to perform the function of this
nginx config.