[#65967404] Feature flag mirror upload and check
Use a class param and hiera to feature flag which environments we want
`govuk_mirrorer` to perform the upload and monitor it using Icinga.

This will allow the update process (which uses a `hosts(5)` entry to run
against the *current* environment) to run everywhere. But the upload process
(which syncs to a common set of offsite machines) to only occur in one
place. `govuk_upload_mirror` will log that the target list was empty.

Depends on gds/deployment#144 which will preserve existing behaviour in
Interim Production.

NB: The beauty of hiera data types and stdlib `validate` means that we no
longer need to cast single strings to arrays. We can just raise an error.
Class params also mean that we have properly namespaced keys in Hiera!