[#65967404] Feature flag the mirror's cronjob
The mirror update process takes a really long time - nearly 10hrs. While it
would be nice to run it in Preview and Staging, so that we can test changes
before they hit production, it's probably more detrimental to run it on all
environments at the same time and also while:

- the mirror requests are being replayed back from prod by Gor
- data sync processes are happening on Preview and Staging

This feature flag will remove the cron by default. The script can still be
run manually with `sudo -u govuk-netstorage …`.

Depends on gds/deployment#147 and gds/deployment#148 being merged first.