[#65806948] Use sudo module from Puppet Forge
Use Saz's 'sudo' module from the Puppet Forge and provide a thin wrapper
for it called 'govuk_sudo'.

This retains our current sudoers file and the Puppet tests from the
'sudo' Puppet module being replaced.

Using the upstream 'sudo' module provides the benefit of being able to
create 'sudoers.d' entries in Hiera.

---

Note that we are using `hiera_hash()` as a workaround for this bug in
Hiera:
https://tickets.puppetlabs.com/browse/HI-118

This bug means that Hiera will not merge hashes from different data
sources when using automatic parameter lookup.

Once this bug is fixed, `hiera_hash()` should be removed and we can rely
on automatic parameter lookup.

See also the note at the very bottom of this document:
http://docs.puppetlabs.com/hiera/1/lookup_types.html#hash-merge