Preserve existing MySQLd PID path
So that we don't *need* to restart MySQLd.

We can "cheat" this branch out by stopping Puppet, transferring the new
`my.cnf` (which is functionally the same), deleting `conf.d/serverid.cnf`,
deploying Puppet, and then letting it run.

It would be desirable, but not essential, to revert this in the future if
our applications cope better with us restarting MySQLd.