[#60945276] Set retry_interval for APT security
We've noticed that the check for packages handled by unattended-upgrades
alerts before the intended 24hr period:
```
[2014-02-24 14:50:54] SERVICE ALERT: monitoring-1.management.production;outstanding security updates;CRITICAL;HARD;24;CRITICAL: 1 security updates need to be applied
[2014-02-24 14:49:37] SERVICE ALERT: monitoring-1.management.production;outstanding security updates;CRITICAL;SOFT;23;CRITICAL: 1 security updates need to be applied
[2014-02-24 14:48:37] SERVICE ALERT: monitoring-1.management.production;outstanding security updates;CRITICAL;SOFT;22;CRITICAL: 1 security updates need to be applied
[2014-02-24 14:47:37] SERVICE ALERT: monitoring-1.management.production;outstanding security updates;CRITICAL;SOFT;21;CRITICAL: 1 security updates need to be applied
```

It looks like this is because the `check_interval` option defines how often
Icinga will check a healthy service. But as soon as it goes into a
non-healthy state subsequent checks will then be scheduled according to the
`retry_interval` option:

http://docs.icinga.org/latest/en/objectdefinitions.html#objectdefinitions-service

Keeping the existing param and value for `check_interval` because it's not
useful to run this check more frequently than `apt-get update` runs.