Update pool of NTP servers.
Some of the pool.ntp.org servers come and go.  This has led to a number
of icinga alerts when that happens.  Using the ubuntu and apple servers
as well as some of the pool ones should help to mitigate this, and
ensure that we always have 2 truechimers.  According to the ntp docs[1]
it's best to have 2n+1 servers where n > 2, so I've removed one of the
pool servers to give us 5.

[1] http://support.ntp.org/bin/view/Support/SelectingOffsiteNTPServers#Section_5.3.3.