[#66132218] Make virus_scan work w/o CLEAN_DIR
A regression was introduced in d1c7ae3 which prevented this script from
being run without a third argument for `CLEAN_DIR`. This broke the
`virus-scan-clean` and `virus-scan-clean-draft` crons on `asset-master-1`
which are intended to periodically re-scan existing files against new virus
definitions. Although nobody noticed the errors in the `machine.email@`
lists.

Allow it to be run without a third argument again by moving the `set -u`
check for unbound variables after the argument checking and usage message. I
could have use `${3:-}` but this approach also allows the usage message to
be printed if you don't give the right number of arguments, which is more
user friendly.