[#65806948] Use fork of hiera-puppet-helper
Switch to our own fork of hiera-puppet-helper because the fix we need in
d6dfc04 hasn't yet been published in the gem.

The specified branch in this fork is equivalent to version 1.0.1 of the
Gem as was previously used, plus d6dfc04 which was cherry-picked.

We can't use the original author's repo with the most recent changes
because the tests currently fail.