Clarify EFG mysql users and fix for puppetlabs/mysql
Previously in 9c8eff3 we were creating two mysql::db with the same
user (efg@%) but two different passwords. What actually happened is
we got two databases and one user. The second password was ignored,
because before creating it we would check for the user's existence and
assume it already existed.

This replicates that behaviour (there is no second password) but
the efg user can access both databases